requirements.txt
datasets
huggingface_hub
pillow
matplotlib
numpy
torch>=2.0.0
transformers>=4.35.0
accelerate
evaluate
tqdm
